pyproject.toml
[tool.poetry]
name = "mcp-mocks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "mcp_mocks" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110"
uvicorn = "^0.29"
requests = "^2.32"
fastmcp = { git = "https://github.com/jlowin/fastmcp.git" }
loguru = "^0.7.3"
ollama = "^0.5.1"
python-dotenv = "^1.1.1"
pydantic-csv = "^0.1.0"
google-generativeai = "^0.8.5"
brooklyn-api-gcp-commons = {git = "https://github.com/dwellci/brooklyn-api-gcp-commons.git", rev = "v1.0.0"}
google-genai = "^1.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
